+++ src/main/java/lotto/constant/Constant.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.constant;
 
 public class Constant {
     // Error Messages
