+++ src/main/java/lotto/machine/KioskStart.java
@@ -1,6 +1,9 @@
-package lotto;
+package lotto.machine;
 
-public class GameStart {
+import lotto.view.LottoKiosk;
+import lotto.repository.Memory;
+
+public class KioskStart {
     public static void main(String[] args) {
         LottoSeller lottoSeller = new LottoSeller();
         Memory memory = new Memory();
